pyproject.toml
[project]
name = "multiagent-news-insight"
version = "0.1.0"
description = "Multi-agent system that collects, analyzes, and summarizes financial news to generate actionable market insights."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.119.0",
    "google-adk[a2a]>=1.16.0",
    "httpx>=0.27.0",
    "litellm>=1.78.5",
    "newsapi-python>=0.2.7",
    "openai>=2.4.0",
    "sentence-transformers>=5.1.1",
    "uvicorn>=0.31.0",
]

[dependency-groups]
dev = [
    "mypy>=1.18.2",
    "ruff>=0.14.1",
]

[tool.ruff]
# 검사할 경로 지정
src = ["."]
# 무시할 디렉토리
exclude = ["venv", ".venv", "build", "dist"]

# 코드 포맷팅 설정 (black 호환)
line-length = 119
indent-width = 4
target-version = "py313"

# 린트 규칙 설정
lint.select = ["E", "F", "I", "UP", "B"]
lint.ignore = ["E501"]  # 필요 시 조정
